+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [0.1.3] - 2020-11-09
+### Added
+- Add support of API EnvelopeNotification
+- Add more tests for increase code coverage
+
+## [0.1.2] - 2020-11-03
+### Fixed
+- Fix RuntimeException Message Body
+
 ## [0.1.1] - 2020-10-23
 ### Fixed
 - Fixed UriResolver URI to https://api.digisign.org